chore: Refactor goal amount meta implimentation.
